feat:移動 StatusUI と ProfileUI のコードを新しいファイルに移動し、構造を整理
