+++ src/ui/StatusUI.js
@@ -1,5 +1,5 @@
-import { formatNumber } from "./utils.js";
-import { API_CONFIG } from "./AppSettings.js";
+import { formatNumber } from "../utils.js";
+import { API_CONFIG } from "../AppSettings.js";
 
 export class StatusUI {
   constructor(rootElement) {
